requirements.txt
# Aplantida ML Training - Requirements
# Python 3.10+ required

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.14.0
tensorflowjs>=4.11.0

# Data handling
pymongo>=4.6.0
pandas>=2.1.3
numpy>=1.26.2
pyarrow>=14.0.1

# Image processing
Pillow>=10.1.0
albumentations>=1.3.1
opencv-python-headless>=4.8.1

# ML utilities
scikit-learn>=1.3.2
timm>=0.9.12

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0

# HTTP & networking
requests>=2.31.0
aiohttp>=3.9.1

# Utilities
tqdm>=4.66.1
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Export tools
onnx>=1.15.0
onnxruntime>=1.16.3

# Testing
pytest>=7.4.3

# Logging
tensorboard>=2.15.1
